functional crawler extracting company information
as a part of work for a non-profit organization i needed information on furniture manufacturers in Serbia. there was a lot of them so a crawler seemed appropriate. i chose to output info in terminal bc i had problems with Unicode codec in the html file and python didnt want to output to a file